test(quick-01): consolidate wall_clock + volume burn_rate tests (#35)
- New burn_rate_wall_clock_test.rs (6 tests): multi-day (7d/30d/90d) +
  very-long-session precision/formatting/arithmetic
- New burn_rate_volume_test.rs (4 tests): high-volume/tiny-cost rounding/
  high-frequency/mixed-size
- Refactored onto shared helpers (apply/update/set_start_time/session_*)
- #[serial] on all (shared process env + config cache)
- Deleted 3 absorbed original files
- 11 burn_rate files -> 1 helper + 4 thematic files; 22 tests preserved 1:1
